+++ CHANGELOG.md
@@ -9,6 +9,10 @@ This project adheres to [Semantic Versioning], with the exception that minor rel
 
 ## [Unreleased]
 
+### Added
+
+- Noise model strengths can now be sampled from a normal distribution ([#329]) ([**@aaronleesander**])
+
 ## [0.4.0] - 2026-02-05
 
 ### Added
@@ -76,6 +80,7 @@ _📚 Refer to the [GitHub Release Notes](https://github.com/munich-quantum-tool
 
 <!-- PR links -->
 
+[#329]: https://github.com/munich-quantum-toolkit/yaqs/pull/329
 [#319]: https://github.com/munich-quantum-toolkit/yaqs/pull/319
 [#316]: https://github.com/munich-quantum-toolkit/yaqs/pull/316
 [#311]: https://github.com/munich-quantum-toolkit/yaqs/pull/311